requirements.txt
# HDD Folder Restructure Tool - Dependencies
# ==========================================

# Google Gemini AI SDK for LLM integration
google-generativeai>=0.8.0

# Load environment variables from .env file
python-dotenv>=1.0.0
tqdm>=4.66.0

